+++ src/sentry/models/integration.py
@@ -1,6 +1,4 @@
 import logging
-from enum import Enum
-from typing import Optional
 
 from django.db import IntegrityError, models
 from django.utils import timezone
@@ -61,72 +59,6 @@ class RepositoryProjectPathConfig(DefaultFieldsModel):
         app_label = "sentry"
         db_table = "sentry_repositoryprojectpathconfig"
         unique_together = (("project", "stack_root"),)
-
-
-class ExternalProviders(Enum):
-    GITHUB = 0
-    GITLAB = 1
-    EMAIL = 100
-    SLACK = 110
-
-
-EXTERNAL_PROVIDERS = {
-    ExternalProviders.GITHUB: "github",
-    ExternalProviders.GITLAB: "gitlab",
-    ExternalProviders.EMAIL: "email",
-    ExternalProviders.SLACK: "slack",
-}
-
-
-def get_provider_name(value: int) -> Optional[str]:
-    return EXTERNAL_PROVIDERS.get(ExternalProviders(value))
-
-
-class ExternalProviderMixin:
-    def get_provider_string(provider_int):
-        return get_provider_name(provider_int) or "unknown"
-
-    def get_provider_enum(provider_str):
-        inv_providers_map = {v: k for k, v in EXTERNAL_PROVIDERS.items()}
-        return inv_providers_map[provider_str].value if inv_providers_map[provider_str] else None
-
-
-class ExternalTeam(DefaultFieldsModel, ExternalProviderMixin):
-    __core__ = False
-
-    team = FlexibleForeignKey("sentry.Team")
-    provider = BoundedPositiveIntegerField(
-        choices=(
-            (ExternalProviders.GITHUB, "github"),
-            (ExternalProviders.GITLAB, "gitlab"),
-        ),
-    )
-    # external_name => the Github/Gitlab team name. Column name is vague to be reused for more external team identities.
-    external_name = models.TextField()
-
-    class Meta:
-        app_label = "sentry"
-        db_table = "sentry_externalteam"
-        unique_together = (("team", "provider", "external_name"),)
-
-
-class ExternalUser(DefaultFieldsModel, ExternalProviderMixin):
-    __core__ = False
-
-    organizationmember = FlexibleForeignKey("sentry.OrganizationMember")
-    provider = BoundedPositiveIntegerField(
-        choices=(
-            (ExternalProviders.GITHUB, "github"),
-            (ExternalProviders.GITLAB, "gitlab"),
-        ),
-    )
-    # external_name => the Github/Gitlab username. Column name is vague to be reused for more external user identities.
-    external_name = models.TextField()
-
-    class Meta:
-        app_label = "sentry"
-        db_table = "sentry_externaluser"
-        unique_together = (("organizationmember", "provider", "external_name"),)
 
 
 class OrganizationIntegration(DefaultFieldsModel):
